Update readme.txt
I updated the code to be able to use 3 data types which are int, boolean, and char. however. The issue with the code is that the return is always an integer. so if i have a code let x = 'a' and then i do return a i will get 65 which is the ascii number.